[BREAKING] Remove FunctionExpression
Remove all of the cases where Glimmer was sticking functions in the
constant pool (which aren't serializable).

The breaking change is that dynamic attributes should be installed by
the didCreateElement manager hook (note: this is *not* the place where
user code is invoked) rather than using the wrapping DSL. Similarly,
dynamic tag name is now modelled by using a getTagName hook on the
component manager.

These hooks should be run in SSR mode.